backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.1
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
apscheduler==3.10.4
redis==5.0.1
python-multipart==0.0.6
cors==1.0.1
pytest==7.4.3
pytest-asyncio==0.21.1